+++ dtslint/ts3.5/ReadonlyArray.ts
@@ -315,3 +315,13 @@ pipe(prns, _.findFirst(predicate))
 pipe(prsns, _.findLast(isString))
 // $ExpectType Option<number>
 pipe(prns, _.findLast(predicate))
+
+//
+// isEmpty
+//
+
+// $ExpectType Either<readonly string[], readonly []>
+pipe(
+  rss,
+  E.fromPredicate(_.isEmpty, (as) => as)
+)