+++ src/These.ts
@@ -675,13 +675,13 @@ export const fromOptionK =
 /**
  * @since 2.11.0
  */
-export const elem = <A>(E: Eq<A>) => (a: A) => (ma: These<unknown, A>): boolean =>
+export const elem = <A>(E: Eq<A>) => (a: A) => <E>(ma: These<E, A>): boolean =>
   isLeft(ma) ? false : E.equals(a, ma.right)
 
 /**
  * @since 2.11.0
  */
-export const exists = <A>(predicate: Predicate<A>) => (ma: These<unknown, A>): boolean =>
+export const exists = <A>(predicate: Predicate<A>) => <E>(ma: These<E, A>): boolean =>
   isLeft(ma) ? false : predicate(ma.right)
 
 /**