+++ CHANGELOG.md
@@ -15,6 +15,27 @@
 **Note**: A feature tagged as Experimental is in a
 high state of flux, you're at risk of it changing without notice.
 
+# 2.11.0-rc.2
+
+- `Array` / `ReadonlyArray`
+  - revert `isOutOfBound` signature change
+  - revert `isEmpty` signature change
+  - revert `size` signature change
+- `Either`
+  - revert `exists` signature change
+  - revert `elem` signature change
+- `These`
+  - revert `exists` signature change
+  - revert `elem` signature change
+- `NonEmptyArray` / `ReadonlyNonEmptyArray`
+  - revert `isOutOfBound` signature change
+- `Set` / `ReadonlySet`
+  - revert `isEmpty` signature change
+  - revert `size` signature change
+- `Map` / `ReadonlyMap`
+  - revert `isEmpty` signature change
+  - revert `size` signature change
+
 # 2.11.0-rc.1
 
 - **Deprecation**